Update deps (#84)
* update halide

* update onnxruntime and DirectML

* update obs for windows build

* update getting onnxruntime source

* fix onnxruntime version directory name

* unzip -q

* change zip -> tgz

* fix

* fix command line

* fix

* back to git clone

* fix build script

+++ scripts/build_mac.sh
@@ -4,13 +4,21 @@ set -xe
 pushd $(dirname $0)/..
 
 OBS_VERSION=$(brew info --json=v2 --cask obs | jq -r .casks[0].version)
-LLVM_VERSION=$(brew info --json=v2 llvm@12 | jq -r .formulae[0].installed[0].version)
+LLVM_VERSION=$(brew info --json=v2 llvm@13 | jq -r .formulae[0].installed[0].version)
 
 echo "Using OBS ${OBS_VERSION}, LLVM ${LLVM_VERSION}"
 
 [ -d deps ] || mkdir deps
-[ -d deps/obs-studio ] && rm -rf deps/obs-studio
-git -C deps clone --single-branch --depth 1 -b ${OBS_VERSION} https://github.com/obsproject/obs-studio.git
+if [ ! -d deps/obs-studio/build ]; then
+  [ -d deps/obs-studio ] && rm -rf deps/obs-studio
+  git -C deps clone --recursive --single-branch --depth 1 -b ${OBS_VERSION} https://github.com/obsproject/obs-studio.git
+
+  pushd deps/obs-studio
+    ./CI/build-macos.sh
+  popd
+fi
+
+
 [ -d build ] && rm -rf build
 mkdir build
 pushd build
@@ -18,12 +26,12 @@ pushd build
   cmake .. \
     -DCMAKE_OSX_ARCHITECTURES="x86_64" \
     -DCMAKE_APPLE_SILICON_PROCESSOR=x86_64 \
-    -DobsLibPath=/Applications/OBS.app/Contents/Frameworks \
-    -DobsIncludePath=$(cd ../deps/obs-studio/libobs; pwd) \
     -DOnnxRuntimePath=$(cd ../deps/onnxruntime; pwd) \
     -DHalide_DIR=$(cd ../deps/Halide; pwd)/lib/cmake/Halide \
     -DHalideHelpers_DIR=$(cd ../deps/Halide; pwd)/lib/cmake/HalideHelpers \
-    -DLLVM_DIR=/usr/local/Cellar/llvm/${LLVM_VERSION}/lib/cmake/llvm
+    -DLLVM_DIR=/usr/local/Cellar/llvm@13/${LLVM_VERSION}/lib/cmake/llvm \
+    -DCMAKE_PREFIX_PATH=$(cd ../deps/obs-studio/build; pwd)
+
   cmake --build . --config Release
   cpack
 popd
